Refactoring cloudcity classes out of OMNT codebase
Since I introduced a new 'cloudcity' package a long time ago, and old networking
code was still in OMNT/cloudcity package, lets move all of that stuff to
cloudcity/networking since that's what it is and that's where it really belongs to.

Then we can do some more refactoring to easily reuse current network sending methods
to send multiple types of requests.
