+++ app/src/main/java/cloudcity/networking/models/MeasurementsModel.java
@@ -1,4 +1,4 @@
-package de.fraunhofer.fokus.OpenMobileNetworkToolkit.cloudCity.models;
+package cloudcity.networking.models;
 
 import android.telephony.CellInfo;
