+++ app/src/main/java/cloudcity/networking/models/NetworkDataModel.java
@@ -1,4 +1,4 @@
-package de.fraunhofer.fokus.OpenMobileNetworkToolkit.cloudCity.models;
+package cloudcity.networking.models;
 
 import com.google.gson.annotations.SerializedName;
